Found a fat bug.
When a user authenticates via LDAP, his/her existing data is replaced.
You might add User x to group B, but when the user logs in, all previously
propagated permissions are replaced by the LDAP app.
I think this is because of the Mirror = True thing I set up.
There was an alternative form of authentication.
I can play around with that, or set it up to work with security groups.