Added note system form with public/private
Still doesn't display, but the HTML form and views for creating a note work as
well as the backend.

We need to improve the alert feedback system for better UX.
Right now we can only send one feedback message at a time, but in this case we
want to send two separate messages. Right now I am embedding the second message
in the first one, but this is sub-optimal.

We also need to add a way for the users to view the PUBLIC NOTES and the event
Admin to view the PUBLIC NOTES AND PRIVATE NOTES (next addition)